feat: implement review management and add large dataset

+++ review-analytics-server/src/main/java/com/github/stasangelov/reviewanalytics/entity/Review.java
@@ -24,7 +24,7 @@ public class Review {
     private LocalDateTime dateCreated;
 
     @Column(name = "integral_rating")
-    private Integer integralRating; // Общий рейтинг
+    private Double integralRating; // Общий рейтинг
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20, nullable = false)
